Simulate example data for fitting GAMs

A tidy reimplementation of the functions implemented in mgcv::gamSim() that can be used to fit GAMs. An new feature is that the sampling distribution can be applied to all the example types.

Usage

data_sim(
  model = "eg1",
  n = 400,
  scale = 2,
  theta = 3,
  dist = c("normal", "poisson", "binary", "negbin", "tweedie"),
  seed = NULL
)

Arguments

model

character; either "egX" where X is an integer 1:7, or the name of a model. See Details for possible options.

n

numeric; the number of observations to simulate.

scale

numeric; the level of noise to use.

theta

numeric; the dispersion parameter \(\theta\) to use. The default is entirely arbitrary, chosen only to provide simulated data that exhibits extra dispersion beyond that assumed by under a Poisson.

dist

character; a sampling distribution for the response variable.

seed

numeric; the seed for the random number generator. Passed to base::set.seed().

Examples

# \dontshow{
set.seed(1)
op <- options(pillar.sigfig = 5, cli.unicode = FALSE)
# }
data_sim("eg1")
#> # A tibble: 400 x 10
#>            y       x0       x1       x2      x3       f      f0     f1        f2
#>        <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>  <dbl>     <dbl>
#>  1  3.3407   0.26551  0.65888  0.85869  0.36653  5.5145 1.4814  3.7350 2.9808e-1
#>  2 -0.075760 0.37212  0.18507  0.034439 0.74139  3.5764 1.8408  1.4479 2.8770e-1
#>  3 10.683    0.57285  0.95438  0.97100  0.93351  8.6926 1.9478  6.7447 8.6114e-5
#>  4  8.7291   0.90821  0.89785  0.74511  0.67321  8.7529 0.56879 6.0237 2.1604e+0
#>  5 14.991    0.20168  0.94370  0.27326  0.70136 16.190  1.1841  6.6021 8.4041e+0
#>  6  7.6709   0.89839  0.72369  0.67711  0.84763  8.0268 0.62765 4.2520 3.1472e+0
#>  7  7.5833   0.94468  0.37036  0.34795  0.70616  8.4353 0.34587 2.0974 5.9920e+0
#>  8  8.5145   0.66080  0.78102  0.94702  0.85881  6.5211 1.7502  4.7685 2.4301e-3
#>  9 10.646    0.62911  0.011150 0.33862  0.44564  9.1904 1.8377  1.0225 6.3302e+0
#> 10  3.7212   0.061786 0.94031  0.031713 0.67663  7.1744 0.38578 6.5576 2.3108e-1
#> # ... with 390 more rows, and 1 more variable: f3 <dbl>
options(op)