Effective degrees of freedom for smooths and GAMs

Extracts the effective degrees of freedom (EDF) for model smooth terms or overall EDF for fitted GAMs

Usage

edf(object, ...)

# S3 method for gam
edf(
  object,
  smooth = NULL,
  type = c("default", "unconditional", "alternative"),
  ...
)

model_edf(object, ..., type = c("default", "unconditional", "alternative"))

Arguments

object

a fitted model from which to extract smooth-specific EDFs.

...

arguments passed to methods.

smooth

character; a vector of smooth terms whose EDFs will be extracted. If NULL, the default, EDFs for all smooths will be returned.

type

character: which type of EDF to return. "default" returns the standard EDF; "unconditional" selects the EDF corrected for smoothness parameter selection, if available; "alternative" returns the alternative formulation for EDF from Wood (2017, pp. 252)

Details

Multiple formulations for the effective degrees of freedom are available. The additional uncertainty due to selection of smoothness parameters can be taken into account when computing the EDF of smooths. This form of the EDF is available with type = "unconditional".

Wood (2017; pp. 252) describes an alternative EDF for the model $$\mathrm{EDF} = 2\mathrm{tr}(\mathbf{F}) - \mathrm{tr}(\mathbf{FF}),$$ where \(\mathrm{tr}\) is the matrix trace and \(\mathbf{F}\) is a matrix mapping un-penalized coefficient estimates to the penalized coefficient estimates. The trace of \(\mathbf{F}\) is effectively the average shrinkage of the coefficients multipled by the number of coefficients (Wood, 2017). Smooth-specific EDFs then are obtained by summing up the relevent elements of \(\mathrm{diag}(2\mathbf{F} - \mathbf{FF})\).

Examples

load_mgcv()
op <- options(cli.unicode = FALSE, digits = 6)
df <- data_sim("eg1", n = 400, seed = 42)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = "REML")

# extract the EDFs for all smooths
edf(m)
#> # A tibble: 4 x 2
#>   smooth   edf
#>   <chr>  <dbl>
#> 1 s(x0)   3.42
#> 2 s(x1)   3.22
#> 3 s(x2)   7.90
#> 4 s(x3)   1.88

# or selected smooths
edf(m, smooth = c("s(x0)", "s(x2)"))
#> # A tibble: 2 x 2
#>   smooth   edf
#>   <chr>  <dbl>
#> 1 s(x0)   3.42
#> 2 s(x2)   7.90

# accounting for smoothness parameter uncertainty
edf(m, type = "unconditional")
#> # A tibble: 4 x 2
#>   smooth   edf
#>   <chr>  <dbl>
#> 1 s(x0)   3.77
#> 2 s(x1)   3.87
#> 3 s(x2)   8.07
#> 4 s(x3)   2.88

# over EDF of the model, including the intercept
model_edf(m)
#> # A tibble: 1 x 2
#>   model   edf
#>   <chr> <dbl>
#> 1 m      17.4

# can get model EDF for multiple models
m2 <- gam(y ~ s(x0) + s(x1) + s(x3), data = df, method = "REML")
model_edf(m, m2)
#> # A tibble: 2 x 2
#>   model   edf
#>   <chr> <dbl>
#> 1 m     17.4 
#> 2 m2     7.58
options(op)