Evaluate a smooth

Evaluate a smooth at a grid of evenly spaced value over the range of the covariate associated with the smooth. Alternatively, a set of points at which the smooth should be evaluated can be supplied.

Usage

evaluate_smooth(object, ...)

# S3 method for gam
evaluate_smooth(
  object,
  smooth,
  n = 100,
  newdata = NULL,
  unconditional = FALSE,
  overall_uncertainty = TRUE,
  dist = 0.1,
  ...
)

# S3 method for gamm
evaluate_smooth(object, ...)

# S3 method for list
evaluate_smooth(object, ...)

Arguments

object

an object of class "gam" or "gamm".

...

arguments passed to other methods.

smooth

character; a single smooth to evaluate.

n

numeric; the number of points over the range of the covariate at which to evaluate the smooth.

newdata

a vector or data frame of points at which to evaluate the smooth.

unconditional

logical; should confidence intervals include the uncertainty due to smoothness selection? If TRUE, the corrected Bayesian covariance matrix will be used.

overall_uncertainty

logical; should the uncertainty in the model constant term be included in the standard error of the evaluate values of the smooth?

dist

numeric; if greater than 0, this is used to determine when a location is too far from data to be plotted when plotting 2-D smooths. The data are scaled into the unit square before deciding what to exclude, and dist is a distance within the unit square. See mgcv::exclude.too.far() for further details.

Value

A data frame, which is of class "evaluated_1d_smooth" or evaluated_2d_smooth, which inherit from classes "evaluated_smooth"and "data.frame".

Details

evaluate_smooth() is deprecated in favour of smooth_estimates(), which provides a cleaner way to evaluate a smooth over a range of covariate values. smooth_estimates() can handle a much wider range of models than evaluate_smooth() is capable of and smooth_estimates() is much easier to extend to handle new smooth types.

Most code that uses evaluate_smooth() should work simply by changing the function call to smooth_estimates(). However, there are some differences:

the newdata argument becomes data

Consider evaluate_smooth() to be soft-deprecated; its use is discouraged and it may be removed at a later date if it becomes difficult to maintain the current functionality, but there are no intentions of removing it from gratia unless that situation arises.

Examples

load_mgcv()
# \dontshow{
set.seed(2)
op <- options(cli.unicode = FALSE, digits = 5)
# }
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
#> Gu & Wahba 4 term additive model
m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

evaluate_smooth(m1, "s(x1)")
#> # A tibble: 100 x 5
#>    smooth by_variable       x1   est    se
#>    <chr>  <fct>          <dbl> <dbl> <dbl>
#>  1 s(x1)  NA          0.000663 -1.94 0.329
#>  2 s(x1)  NA          0.0108   -1.93 0.312
#>  3 s(x1)  NA          0.0208   -1.92 0.296
#>  4 s(x1)  NA          0.0309   -1.90 0.281
#>  5 s(x1)  NA          0.0410   -1.89 0.267
#>  6 s(x1)  NA          0.0511   -1.87 0.255
#>  7 s(x1)  NA          0.0612   -1.86 0.243
#>  8 s(x1)  NA          0.0713   -1.84 0.234
#>  9 s(x1)  NA          0.0814   -1.83 0.225
#> 10 s(x1)  NA          0.0914   -1.81 0.218
#> # ... with 90 more rows

## 2d example
set.seed(2)
dat <- gamSim(2, n = 1000, dist = "normal", scale = 1)
#> Bivariate smoothing example
m2 <- gam(y ~ s(x, z, k = 30), data = dat$data, method = "REML")

evaluate_smooth(m2, "s(x,z)", n = 100)
#> # A tibble: 10,000 x 6
#>    smooth by_variable        x        z    est    se
#>    <chr>  <fct>          <dbl>    <dbl>  <dbl> <dbl>
#>  1 s(x,z) NA          0.000663 0.000762 0.0881 0.189
#>  2 s(x,z) NA          0.0108   0.000762 0.0922 0.186
#>  3 s(x,z) NA          0.0208   0.000762 0.0965 0.183
#>  4 s(x,z) NA          0.0309   0.000762 0.101  0.180
#>  5 s(x,z) NA          0.0410   0.000762 0.105  0.176
#>  6 s(x,z) NA          0.0511   0.000762 0.110  0.173
#>  7 s(x,z) NA          0.0612   0.000762 0.115  0.170
#>  8 s(x,z) NA          0.0713   0.000762 0.120  0.166
#>  9 s(x,z) NA          0.0814   0.000762 0.125  0.163
#> 10 s(x,z) NA          0.0914   0.000762 0.130  0.160
#> # ... with 9,990 more rows
options(op)