Create a sequence of evenly-spaced values

For a continuous vector x, evenly and seq_min_max() create a sequence of n evenly-spaced values over the range lower -- upper. By default, lower is defined as min(x) and upper as max(x), excluding NAs. For a factor x, the function returns levels(x).

Usage

evenly(x, n = 100, by = NULL, lower = NULL, upper = NULL)

seq_min_max(x, n, by = NULL, lower = NULL, upper = NULL)

Arguments

x

numeric; vector over which evenly-spaced values are returned

n

numeric; the number of evenly-spaced values to return. A default of 100 is used for convenience as that what is typically used when evaluating a smooth.

by

numeric; the increment of the sequence. If specified, argument n is ignored and the sequence returned will be from min(x) to max(x) in increments of by.

lower

numeric; the lower bound of the interval.

upper

numeric; the upper bound of the interval.

Value

A numeric vector of length n.

Examples

set.seed(1)
x <- rnorm(10)
n <- 10L
evenly(x, n = n)
#>  [1] -0.83562861 -0.56552757 -0.29542652 -0.02532547  0.24477557  0.51487662
#>  [7]  0.78497766  1.05507871  1.32517976  1.59528080