Values for rug plot in nested form

Extracts original data for smooth terms, formats them in long/tidy format, then nests the data column(s) such that the result is a nested data frame with one row per smooth.

Usage

nested_rug_values(object, terms = NULL, data = NULL)

Arguments

object

a fitted GAM model

terms

a vector of terms to include original data for. Passed to argument terms of mgcv::predict.gam()].

data

optional data frame

Value

A nested tibble (data frame) with one row per smooth term.