Extract and tidy penalty matrices

Extract and tidy penalty matrices

penalty(object, ...)

# S3 method for gam
penalty(object, smooth = NULL, rescale = FALSE, ...)

# S3 method for mgcv.smooth
penalty(object, rescale = FALSE, ...)

# S3 method for tensor.smooth
penalty(object, margins = FALSE, ...)

# S3 method for t2.smooth
penalty(object, margins = FALSE, ...)

Arguments

object
a fitted GAM or a smooth.

...
additional arguments passed to methods.

smooth
character; vector of smooths to extract penalty matrices for. If NULL, penalty matrices for all smooths in object are extracted.

rescale
logical; by default, mgcv will scale the penalty matrix for better performance in mgcv::gamm(). If rescale is TRUE, this scaling will be undone to put the penalty matrix back on the original scale.

margins
logical; extract the penalty matrices for the tensor product or the marginal smooths of the tensor product?

Value

A 'tibble' (data frame) of class penalty_df inheriting from tbl_df, with the following components:

smooth - character; the label mgcv uses to refer to the smooth,

type - character; the type of smooth,

penalty - character; the label for the specific penalty. Some smooths have multiple penalty matrices, so the penalty component identifies the particular penalty matrix and uses the labelling that mgcv uses internally,

row - character; a label of the form fn where n is an integer for the nth basis function, referencing the columns of the penalty matrix,

col - character; a label of the form fn where n is an integer for the nth basis function, referencing the columns of the penalty matrix,

value - double; the value of the penalty matrix for the combination of row and col,

Note

The print() method uses base::zapsmall() to turn very small numbers into 0s for display purposes only; the underlying values of the penalty matrix or matrices are not changed.

Author

Gavin L. Simpson

Examples

# \dontshow{
op <- options(cli.unicode = FALSE, digits = 5)
# }
load_mgcv()
dat <- data_sim("eg4", n = 400, seed = 42)
m <- gam(y ~ s(x0) + s(x1) + s(x2, by = fac),
         data = dat, method = "REML")
penalty(m)

#> # A tibble: 405 x 6
#>    smooth type  penalty row   col   value
#>    <chr>  <chr> <chr>   <chr> <chr> <dbl>
#>  1 s(x0)  TPRS  s(x0)   f1    f1     9.81
#>  2 s(x0)  TPRS  s(x0)   f1    f2    -1.45
#>  3 s(x0)  TPRS  s(x0)   f1    f3    -5   
#>  4 s(x0)  TPRS  s(x0)   f1    f4    -1.34
#>  5 s(x0)  TPRS  s(x0)   f1    f5    -6.24
#>  6 s(x0)  TPRS  s(x0)   f1    f6     3.90
#>  7 s(x0)  TPRS  s(x0)   f1    f7    -7.74
#>  8 s(x0)  TPRS  s(x0)   f1    f8    -1.79
#>  9 s(x0)  TPRS  s(x0)   f1    f9     0   
#> 10 s(x0)  TPRS  s(x0)   f2    f1    -1.45
#> # ... with 395 more rows

# for a specific smooth
penalty(m, smooth = "s(x2):fac1")

#> # A tibble: 81 x 6
#>    smooth     type  penalty    row   col   value
#>    <chr>      <chr> <chr>      <chr> <chr> <dbl>
#>  1 s(x2):fac1 TPRS  s(x2):fac1 f1    f1     9.94
#>  2 s(x2):fac1 TPRS  s(x2):fac1 f1    f2     1.40
#>  3 s(x2):fac1 TPRS  s(x2):fac1 f1    f3     4.78
#>  4 s(x2):fac1 TPRS  s(x2):fac1 f1    f4     1.13
#>  5 s(x2):fac1 TPRS  s(x2):fac1 f1    f5     5.84
#>  6 s(x2):fac1 TPRS  s(x2):fac1 f1    f6    -4.06
#>  7 s(x2):fac1 TPRS  s(x2):fac1 f1    f7     8.60
#>  8 s(x2):fac1 TPRS  s(x2):fac1 f1    f8     1.44
#>  9 s(x2):fac1 TPRS  s(x2):fac1 f1    f9     0   
#> 10 s(x2):fac1 TPRS  s(x2):fac1 f2    f1     1.40
#> # ... with 71 more rows

options(op)