Draw samples from the posterior distribution of an estimated model

Draw samples from the posterior distribution of an estimated model

posterior_samples(model, ...)

# S3 method for gam
posterior_samples(
  model,
  n,
  newdata,
  seed,
  scale = c("response", "linear_predictor"),
  freq = FALSE,
  unconditional = FALSE,
  weights = NULL,
  ncores = 1L,
  ...
)

Arguments

model
a fitted model of the supported types

...
arguments passed to other methods

n
numeric; the number of posterior samples to return.

newdata
data frame; new observations at which the posterior draws from the model should be evaluated. If not supplied, the data used to fit the model will be used for newdata, if available in model.

seed
numeric; a random seed for the simulations.

scale
character;

freq
logical; TRUE to return the frequentist covariance matrix of the parameter estimators, FALSE to return the Bayesian posterior covariance matrix of the parameters.

unconditional
logical; if TRUE (and freq == FALSE) then the Bayesian smoothing parameter uncertainty corrected covariance matrix is returned, if available.

weights
numeric; a vector of prior weights. If newdata is null then defaults to object[["prior.weights"]], otherwise a vector of ones.

ncores
number of cores for generating random variables from a multivariate normal distribution. Passed to mvnfast::rmvn(). Parallelization will take place only if OpenMP is supported (but appears to work on Windows with current R).

Value

A tibble (data frame) with 3 columns containing the posterior predicted values in long format. The columns are

row (integer) the row of newdata that each posterior draw relates to,

draw (integer) an index, in range 1:n, indicating which draw each row relates to,

response (numeric) the predicted response for the indicated row of newdata.