Default diverging red-blue colour palette for partial effects

Default diverging red-blue colour palette for partial effects

Usage

scale_fill_partial_effect(
  name = "Partial effect",
  ...,
  na.value = "grey50",
  guide = "colourbar",
  direction = -1
)

Arguments

name

The name of the scale. Used as the legend title. If NULL the legend title will be omitted.

...

arguments passed to ggplot2::continuous_scale().

na.value

Missing values will be replaced with this value (colour).

guide

A function used to create a guide or its name. See ggplot2::guides() for more information.

direction

Sets the order of colours in the scale. If 1, the default, colours are as output by RColorBrewer::brewer.pal(). If -1, the order of colours is reversed.