Simulate from the posterior distribution of a GAM

Simulations from the posterior distribution of a fitted GAM model involve making random draws from a multivariate normal with mean vector equal to the estimated model coefficients and covariance matrix equal to the covariance matrix of the coefficients.

# S3 method for gam
simulate(object, nsim = 1, seed = NULL, newdata = NULL,
  freq = FALSE, unconditional = FALSE, ...)

# S3 method for gamm
simulate(object, nsim = 1, seed = NULL,
  newdata = NULL, freq = FALSE, unconditional = FALSE, ...)

# S3 method for scam
simulate(object, nsim = 1, seed = NULL,
  newdata = NULL, freq = FALSE, parametrized = TRUE, ...)

Arguments

object
a fitted GAM, typically the result of a call to gam() or gamm().

nsim
numeric; the number of posterior simulations to return.

seed
numeric; a random seed for the simulations.

newdata
data frame; new observations at which the posterior draws from the model should be evaluated. If not supplied, the data used to fit the model will be used for newdata, if available in object.

freq
logical; TRUE to return the frequentist covariance matrix of the parameter estimators, FALSE to return the Bayesian posterior covariance matrix of the parameters.

unconditional
logical; if TRUE (and freq == FALSE) then the Bayesian smoothing parameter uncertainty corrected covariance matrix is returned, if available.

...
arguments passed to methods

parametrized
logical; use parametrized coefficients and covariance matrix, which respect the linear inequality constraints of the model.

Value

(Currently) A matrix with nsim columns.

Examples

library("mgcv")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

sims <- simulate(m1, nsim = 5, seed = 42)
head(sims)

#>        [,1]      [,2]      [,3]      [,4]      [,5]
#> 1  8.470111  8.502529  9.243253  8.583744  8.888660
#> 2  7.135076  7.166169  7.436497  7.447073  8.036909
#> 3  3.632465  2.544824  3.083198  2.639840  2.612815
#> 4 10.686483 10.415981 11.873220 10.799256 10.865970
#> 5 14.077964 14.273323 14.646527 14.000525 13.707774
#> 6  6.300381  6.205313  6.094017  5.697613  5.943234