Determine the type of smooth and return it n a human readable form

Determine the type of smooth and return it n a human readable form

Usage

smooth_type(smooth)

# Default S3 method
smooth_type(smooth)

# S3 method for class 'tprs.smooth'
smooth_type(smooth)

# S3 method for class 'ts.smooth'
smooth_type(smooth)

# S3 method for class 'cr.smooth'
smooth_type(smooth)

# S3 method for class 'cs.smooth'
smooth_type(smooth)

# S3 method for class 'cyclic.smooth'
smooth_type(smooth)

# S3 method for class 'pspline.smooth'
smooth_type(smooth)

# S3 method for class 'cpspline.smooth'
smooth_type(smooth)

# S3 method for class 'Bspline.smooth'
smooth_type(smooth)

# S3 method for class 'duchon.spline'
smooth_type(smooth)

# S3 method for class 'fs.interaction'
smooth_type(smooth)

# S3 method for class 'sz.interaction'
smooth_type(smooth)

# S3 method for class 'gp.smooth'
smooth_type(smooth)

# S3 method for class 'mrf.smooth'
smooth_type(smooth)

# S3 method for class 'random.effect'
smooth_type(smooth)

# S3 method for class 'sw'
smooth_type(smooth)

# S3 method for class 'sf'
smooth_type(smooth)

# S3 method for class 'soap.film'
smooth_type(smooth)

# S3 method for class 't2.smooth'
smooth_type(smooth)

# S3 method for class 'sos.smooth'
smooth_type(smooth)

# S3 method for class 'tensor.smooth'
smooth_type(smooth)

# S3 method for class 'mpi.smooth'
smooth_type(smooth)

# S3 method for class 'mpd.smooth'
smooth_type(smooth)

# S3 method for class 'cx.smooth'
smooth_type(smooth)

# S3 method for class 'cv.smooth'
smooth_type(smooth)

# S3 method for class 'micx.smooth'
smooth_type(smooth)

# S3 method for class 'micv.smooth'
smooth_type(smooth)

# S3 method for class 'mdcx.smooth'
smooth_type(smooth)

# S3 method for class 'mdcv.smooth'
smooth_type(smooth)

# S3 method for class 'miso.smooth'
smooth_type(smooth)

# S3 method for class 'mifo.smooth'
smooth_type(smooth)

Arguments

smooth

an object inheriting from class mgcv.smooth.