Variance components of smooths from smoothness estimates

A wrapper to mgcv::gam.vcomp() which returns the smoothing parameters expressed as variance components.

Usage

variance_comp(object, ...)

# S3 method for class 'gam'
variance_comp(object, rescale = TRUE, coverage = 0.95, ...)

Arguments

object

an R object. Currently only models fitted by mgcv::gam() or mgcv::bam() are supported.

...

arguments passed to other methods

rescale

logical; for numerical stability reasons the penalty matrices of smooths are rescaled before fitting. If rescale = TRUE, this rescaling is undone, resulting in variance components that are on their original scale. This is needed if comparing with other mixed model software, such as lmer().

coverage

numeric; a value between 0 and 1 indicating the (approximate) coverage of the confidence interval that is returned.

Details

This function is a wrapper to mgcv::gam.vcomp() which performs three additional services

it suppresses the annoying text output that mgcv::gam.vcomp() prints to the terminal,

returns the variance of each smooth as well as the standard deviation, and

returns the variance components as a tibble.