Basis expansions for smooths

Creates a basis expansion from a definition of a smoother using the syntax of mgcv's smooths via mgcv::s()., mgcv::te(), mgcv::ti(), and mgcv::t2().

basis(smooth, data, knots = NULL, constraints = FALSE, ...)

Arguments

smooth
a smooth specification, the result of a call to one of mgcv::s()., mgcv::te(), mgcv::ti(), or mgcv::t2().

data
a data frame containing the variables used in smooth.

knots
a list or data frame with named components containing knots locations. Names must match the covariates for which the basis is required. See mgcv::smoothCon().

constraints
logical; should identifiability constraints be applied to the smooth basis. See argument absorb.cons in mgcv::smoothCon().

...
other arguments passed to mgcv::smoothCon().

Value

A tibble.

Author

Gavin L. Simpson

Examples

load_mgcv()
# \dontshow{
set.seed(42)
op <- options(digits = 3, cli.unicode = FALSE)
# }
df <- gamSim(4, n = 400, verbose = FALSE)

bf <- basis(s(x0), data = df)
bf <- basis(s(x2, by = fac, bs = 'bs'), data = df, constraints = TRUE)
options(op)