Select smooths based on user's choices

Given a vector indexing the smooths of a GAM, returns a logical vector selecting the requested smooths.

check_user_select_smooths(smooths, select = NULL, partial_match = FALSE)

Arguments

smooths
character; a vector of smooth labels.

select
numeric, logical, or character vector of selected smooths.

partial_match
logical; in the case of character select, should select match partially against smooths? If partial_match = TRUE, select must only be a single string, a character vector of length 1.

Value

A logical vector the same length as length(smooths) indicating which smooths have been selected.