Point-wise and simultaneous confidence intervals for derivatives of smooths

Calculates point-wise confidence or simultaneous intervals for the first derivatives of smooth terms in a fitted GAM.

# S3 method for fderiv
confint(
  object,
  parm,
  level = 0.95,
  type = c("confidence", "simultaneous"),
  nsim = 10000,
  ncores = 1L,
  ...
)

Arguments

object

an object of class "fderiv" containing the estimated derivatives.

parm

which parameters (smooth terms) are to be given intervals as a vector of terms. If missing, all parameters are considered.

level

numeric, 0 < level < 1; the confidence level of the point-wise or simultaneous interval. The default is 0.95 for a 95% interval.

type

character; the type of interval to compute. One of "confidence" for point-wise intervals, or "simultaneous" for simultaneous intervals.

nsim

integer; the number of simulations used in computing the simultaneous intervals.

ncores

number of cores for generating random variables from a multivariate normal distribution. Passed to mvnfast::rmvn(). Parallelization will take place only if OpenMP is supported (but appears to work on Windows with current R).

...

additional arguments for methods

Value

a data frame with components:

term; factor indicating to which term each row relates,

lower; lower limit of the confidence or simultaneous interval,

est; estimated derivative

upper; upper limit of the confidence or simultaneous interval.

Author

Gavin L. Simpson

Examples

load_mgcv()
# \dontshow{
set.seed(2)
op <- options(digits = 3, cli.unicode = FALSE)
# }
dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
#> Gu & Wahba 4 term additive model
mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## first derivatives of all smooths...
fd <- fderiv(mod)
#> Warning: `fderiv()` was deprecated in gratia 0.7.0.
#> Please use `derivatives()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.

## point-wise interval
ci <- confint(fd, type = "confidence")
head(ci)
#>    term lower  est upper
#> 1 s(x0)  1.22 6.02  10.8
#> 2 s(x0)  1.22 6.02  10.8
#> 3 s(x0)  1.22 6.02  10.8
#> 4 s(x0)  1.23 6.01  10.8
#> 5 s(x0)  1.25 6.01  10.8
#> 6 s(x0)  1.27 6.01  10.7

## simultaneous interval for smooth term of x1
# \dontshow{
set.seed(42)
# }
x1.sint <- confint(fd, parm = "x1", type = "simultaneous", nsim = 2500)
head(x1.sint)
#>    term  lower  est upper
#> 1 s(x1) -0.483 4.70  9.88
#> 2 s(x1) -0.482 4.70  9.88
#> 3 s(x1) -0.478 4.70  9.87
#> 4 s(x1) -0.469 4.70  9.86
#> 5 s(x1) -0.452 4.70  9.84
#> 6 s(x1) -0.427 4.69  9.81
options(op)