Point-wise and simultaneous confidence intervals for derivatives of smooths

Calculates point-wise confidence or simultaneous intervals for the first derivatives of smooth terms in a fitted GAM.

# S3 method for fderiv
confint(object, parm, level = 0.95,
  type = c("confidence", "simultaneous"), nsim = 10000, ...)

Arguments

object
an object of class "fderiv" containing the estimated derivatives.

parm
which parameters (smooth terms) are to be given intervals as a vector of terms. If missing, all parameters are considered.

level
numeric, 0 < level < 1; the confidence level of the point-wise or simultaneous interval. The default is 0.95 for a 95 interval.

type
character; the type of interval to compute. One of "confidence" for point-wise intervals, or "simultaneous" for simultaneous intervals.

nsim
integer; the number of simulations used in computing the simultaneous intervals.

...
additional arguments for methods

Value

a data frame with components:

term; factor indicating to which term each row relates,

lower; lower limit of the confidence or simultaneous interval,

est; estimated derivative

upper; upper limit of the confidence or simultaneous interval.

Examples

library("mgcv")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## first derivatives of all smooths...
fd <- fderiv(mod)

## point-wise interval
ci <- confint(fd, type = "confidence")
head(ci)

#>    term      lower      est    upper
#> 1 s(x0) -0.8496033 4.112256 9.074116
#> 2 s(x0) -0.8489453 4.112287 9.073519
#> 3 s(x0) -0.8448850 4.112468 9.069821
#> 4 s(x0) -0.8329611 4.112988 9.058936
#> 5 s(x0) -0.8108549 4.113933 9.038721
#> 6 s(x0) -0.7769722 4.115360 9.007693

## simultaneous interval for smooth term of x1
set.seed(42)
x1.sint <- confint(fd, parm = "x1", type = "simultaneous", nsim = 1000)
head(x1.sint)

#>    term     lower      est    upper
#> 1 s(x1) -5.383827 1.423165 8.230157
#> 2 s(x1) -5.382662 1.423211 8.229084
#> 3 s(x1) -5.377287 1.423425 8.224137
#> 4 s(x1) -5.366493 1.423860 8.214213
#> 5 s(x1) -5.348794 1.424591 8.197975
#> 6 s(x1) -5.322803 1.425697 8.174198