Point-wise and simultaneous confidence intervals for smooths

Calculates point-wise confidence or simultaneous intervals for the smooth terms of a fitted GAM.

# S3 method for gam
confint(object, parm, level = 0.95, newdata = NULL,
  n = 200, type = c("confidence", "simultaneous"), nsim = 10000,
  shift = FALSE, transform = FALSE, unconditional = FALSE, ...)

# S3 method for gamm
confint(object, ...)

Arguments

object
an object of class "gam" or "gamm".

parm
which parameters (smooth terms) are to be given intervals as a vector of terms. If missing, all parameters are considered, although this is not currently implemented.

level
numeric, 0 < level < 1; the confidence level of the point-wise or simultaneous interval. The default is 0.95 for a 95% interval.

newdata
data frame; containing new values of the covariates used in the model fit. The selected smooth(s) wil be evaluated at the supplied values.

n
numeric; the number of points to evaluate smooths at.

type
character; the type of interval to compute. One of "confidence" for point-wise intervals, or "simultaneous" for simultaneous intervals.

nsim
integer; the number of simulations used in computing the simultaneous intervals.

shift
logical; should the constant term be add to the smooth?

transform
logical; should the smooth be evaluated on a transformed scale? For generalised models, this involves applying the inverse of the link function used to fit the model. Alternatively, the name of, or an actual, function can be supplied to transform the smooth and it's confidence interval.

unconditional
logical; if TRUE (and freq == FALSE) then the Bayesian smoothing parameter uncertainty corrected covariance matrix is returned, if available.

...
additional arguments for methods

Value

a data frame with components:

term; factor indicating to which term each row relates,

x; the vector of values at which the smooth was evaluated,

lower; lower limit of the confidence or simultaneous interval,

est; estimated value of the smooth

upper; upper limit of the confidence or simultaneous interval,

crit; critical value for the 100 * level

Examples

library("mgcv")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
##'
## point-wise interval
ci <- confint(mod, parm = "x1", type = "confidence")
head(ci)

#> # A tibble: 6 x 8
#>   smooth by_variable       x1   est    se  crit lower upper
#>   <chr>  <fct>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 s(x1)  NA          0.000663  5.89 0.329  1.96  5.24  6.53
#> 2 s(x1)  NA          0.00568   5.90 0.320  1.96  5.27  6.52
#> 3 s(x1)  NA          0.0107    5.90 0.312  1.96  5.29  6.51
#> 4 s(x1)  NA          0.0157    5.91 0.304  1.96  5.31  6.51
#> 5 s(x1)  NA          0.0207    5.92 0.296  1.96  5.34  6.50
#> 6 s(x1)  NA          0.0258    5.92 0.288  1.96  5.36  6.49

##'
## simultaneous interval for smooth term of x1
set.seed(42)
si <- confint(mod, parm = "x1", type = "simultaneous", nsim = 100)
head(si)

#> # A tibble: 6 x 8
#>   smooth by_variable       x1   est    se  crit lower upper
#>   <chr>  <fct>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 s(x1)  NA          0.000663  5.89 0.329  2.26  5.15  6.63
#> 2 s(x1)  NA          0.00568   5.90 0.320  2.26  5.17  6.62
#> 3 s(x1)  NA          0.0107    5.90 0.312  2.26  5.20  6.61
#> 4 s(x1)  NA          0.0157    5.91 0.304  2.26  5.22  6.60
#> 5 s(x1)  NA          0.0207    5.92 0.296  2.26  5.25  6.59
#> 6 s(x1)  NA          0.0258    5.92 0.288  2.26  5.27  6.58