Simulate example data for fitting GAMs

A tidy reimplementation of the functions implemented in mgcv::gamSim() that can be used to fit GAMs. An new feature is that the sampling distribution can be applied to all the example types.

Usage

data_sim(
  model = "eg1",
  n = 400,
  scale = 2,
  theta = 3,
  dist = c("normal", "poisson", "binary", "negbin", "tweedie"),
  seed = NULL
)

Arguments

model

character; either "egX" where X is an integer 1:7, or the name of a model. See Details for possible options.

n

numeric; the number of observations to simulate.

scale

numeric; the level of noise to use.

theta

numeric; the dispersion parameter \(\theta\) to use. The default is entirely arbitrary, chosen only to provide simulated data that exhibits extra dispersion beyond that assumed by under a Poisson.

dist

character; a sampling distribution for the response variable.

seed

numeric; the seed for the random number generator. Passed to base::set.seed().

Examples

# \dontshow{
set.seed(1)
op <- options(digits = 5, cli.unicode = FALSE)
# }
data_sim("eg1")
#> # A tibble: 400 x 10
#>          y     x0     x1     x2    x3     f    f0    f1        f2    f3
#>      <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl>
#>  1  3.34   0.266  0.659  0.859  0.367  5.51 1.48   3.74 0.298         0
#>  2 -0.0758 0.372  0.185  0.0344 0.741  3.58 1.84   1.45 0.288         0
#>  3 10.7    0.573  0.954  0.971  0.934  8.69 1.95   6.74 0.0000861     0
#>  4  8.73   0.908  0.898  0.745  0.673  8.75 0.569  6.02 2.16          0
#>  5 15.0    0.202  0.944  0.273  0.701 16.2  1.18   6.60 8.40          0
#>  6  7.67   0.898  0.724  0.677  0.848  8.03 0.628  4.25 3.15          0
#>  7  7.58   0.945  0.370  0.348  0.706  8.44 0.346  2.10 5.99          0
#>  8  8.51   0.661  0.781  0.947  0.859  6.52 1.75   4.77 0.00243       0
#>  9 10.6    0.629  0.0111 0.339  0.446  9.19 1.84   1.02 6.33          0
#> 10  3.72   0.0618 0.940  0.0317 0.677  7.17 0.386  6.56 0.231         0
#> # ... with 390 more rows
options(op)