Prepare a data slice through covariates

Prepare a data slice through covariates

data_slice(object, ...)

# S3 method for default
data_slice(object, ...)

# S3 method for gam
data_slice(object, var1, var2, var3 = NULL, var4 = NULL,
  data = NULL, n = 50, offset = NULL, ...)

Arguments

object
an R model object.

...
arguments passed to other methods.

var1
character;

var2
character;

var3
character; ignored currently.

var4
character; ignored currently.

data
a 1-row data frame or tibble containing values for variables in the fitted model that are not varying in the slice.

n
numeric; the number of values to create for each of var1 and var2in the slice.

offset
numeric; value to use for an offset term in the model.