Derivatives of estimated smooths via finite differences

Derivatives of estimated smooths via finite differences

derivatives(object, ...)

# S3 method for default
derivatives(object, ...)

# S3 method for gamm
derivatives(object, ...)

# S3 method for gam
derivatives(object, term, newdata, order = 1L,
  type = c("forward", "backward", "central"), n = 200, eps = 1e-07,
  unconditional = FALSE, frequentist = FALSE, offset = NULL, ...)

Arguments

object
an R object to compute derivatives for.

...
arguments passed to other methods.

term
character; vector of one or more smooth terms for which derivatives are required. If missing, derivatives for all smooth terms will be returned. Can be a partial match to a smooth term.

newdata
a data frame containing the values of the model covariates at which to evaluate the first derivatives of the smooths.

order
numeric; the order of derivative.

type
character; the type of finite difference used. One of "forward", "backward", or "central".

n
numeric; the number of points to evaluate the derivative at.

eps
numeric; the finite difference.

unconditional
logical; use smoothness selection-corrected Bayesian covariance matrix?

frequentist
logical; use the frequentist covariance matrix?

offset
numeric; a value to use for any offset term

Examples

library("mgcv")
set.seed(42)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2, verbose = FALSE)
mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## first derivative of all smooths
derivatives(mod)

#> # A tibble: 800 x 5
#>    smooth var       data derivative    se
#>    <chr>  <chr>    <dbl>      <dbl> <dbl>
#>  1 s(x0)  x0    0.000239       7.41  3.33
#>  2 s(x0)  x0    0.00525        7.41  3.33
#>  3 s(x0)  x0    0.0103         7.40  3.33
#>  4 s(x0)  x0    0.0153         7.40  3.31
#>  5 s(x0)  x0    0.0203         7.39  3.30
#>  6 s(x0)  x0    0.0253         7.38  3.27
#>  7 s(x0)  x0    0.0303         7.36  3.24
#>  8 s(x0)  x0    0.0353         7.34  3.20
#>  9 s(x0)  x0    0.0403         7.32  3.15
#> 10 s(x0)  x0    0.0453         7.29  3.10
#> # ... with 790 more rows

## second derivative of smooth of x2 only using central finite difference
derivatives(mod, "s(x2)", order = 2, type = "central")

#> # A tibble: 200 x 5
#>    smooth var      data derivative    se
#>    <chr>  <chr>   <dbl>      <dbl> <dbl>
#>  1 s(x2)  x2    0.00359      -9.06 12.0 
#>  2 s(x2)  x2    0.00860      20.1  12.8 
#>  3 s(x2)  x2    0.0136       -1.64  6.94
#>  4 s(x2)  x2    0.0186       19.5  18.6 
#>  5 s(x2)  x2    0.0236       21.7  29.9 
#>  6 s(x2)  x2    0.0286       34.8  36.2 
#>  7 s(x2)  x2    0.0336       66.7  54.7 
#>  8 s(x2)  x2    0.0386       66.1  66.1 
#>  9 s(x2)  x2    0.0436       90.1  80.3 
#> 10 s(x2)  x2    0.0486       83.2  86.5 
#> # ... with 190 more rows