Derivatives of estimated smooths via finite differences

Derivatives of estimated smooths via finite differences

Usage

derivatives(object, ...)

# S3 method for default
derivatives(object, ...)

# S3 method for gamm
derivatives(object, ...)

# S3 method for gam
derivatives(
  object,
  term,
  data = newdata,
  order = 1L,
  type = c("forward", "backward", "central"),
  n = 200,
  eps = 1e-07,
  interval = c("confidence", "simultaneous"),
  n_sim = 10000,
  level = 0.95,
  unconditional = FALSE,
  frequentist = FALSE,
  offset = NULL,
  ncores = 1,
  partial_match = FALSE,
  ...,
  newdata = NULL
)

Arguments

object

an R object to compute derivatives for.

...

arguments passed to other methods.

term

character; vector of one or more smooth terms for which derivatives are required. If missing, derivatives for all smooth terms will be returned. Can be a partial match to a smooth term; see argument partial_match below.

data

a data frame containing the values of the model covariates at which to evaluate the first derivatives of the smooths.

order

numeric; the order of derivative.

type

character; the type of finite difference used. One of "forward", "backward", or "central".

n

numeric; the number of points to evaluate the derivative at.

eps

numeric; the finite difference.

interval

character; the type of interval to compute. One of "confidence" for point-wise intervals, or "simultaneous" for simultaneous intervals.

n_sim

integer; the number of simulations used in computing the simultaneous intervals.

level

numeric; 0 < level < 1; the confidence level of the point-wise or simultaneous interval. The default is 0.95 for a 95% interval.

unconditional

logical; use smoothness selection-corrected Bayesian covariance matrix?

frequentist

logical; use the frequentist covariance matrix?

offset

numeric; a value to use for any offset term

ncores

number of cores for generating random variables from a multivariate normal distribution. Passed to mvnfast::rmvn(). Parallelization will take place only if OpenMP is supported (but appears to work on Windows with current R).

partial_match

logical; should smooths be selected by partial matches with term? If TRUE, term can only be a single string to match against.

newdata

Deprecated: use data instead.

Value

A tibble, currently with the following variables:

smooth: the smooth each row refers to,

var: the name of the variable involved in the smooth,

data: values of var at which the derivative was evaluated,

derivative: the estimated derivative,

se: the standard error of the estimated derivative,

crit: the critical value such that derivative ± (crit * se) gives the upper and lower bounds of the requested confidence or simultaneous interval (given level),

lower: the lower bound of the confidence or simultaneous interval,

upper: the upper bound of the confidence or simultaneous interval.

Note

derivatives() will ignore any random effect smooths it encounters in object.

Author

Gavin L. Simpson

Examples

load_mgcv()
# \dontshow{
op <- options(pillar.sigfig = 3, cli.unicode = FALSE)
# }
dat <- data_sim("eg1", n = 400, dist = "normal", scale = 2, seed = 42)
mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## first derivatives of all smooths using central finite differences
derivatives(mod, type = "central")
#> # A tibble: 800 x 10
#>    smooth var   by_var fs_var     data derivative    se  crit lower upper
#>    <chr>  <chr> <chr>  <chr>     <dbl>      <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 s(x0)  x0    NA     NA     0.000239       7.41  3.33  1.96 0.874  13.9
#>  2 s(x0)  x0    NA     NA     0.00525        7.41  3.33  1.96 0.875  13.9
#>  3 s(x0)  x0    NA     NA     0.0103         7.40  3.33  1.96 0.884  13.9
#>  4 s(x0)  x0    NA     NA     0.0153         7.40  3.31  1.96 0.902  13.9
#>  5 s(x0)  x0    NA     NA     0.0203         7.39  3.30  1.96 0.929  13.8
#>  6 s(x0)  x0    NA     NA     0.0253         7.38  3.27  1.96 0.965  13.8
#>  7 s(x0)  x0    NA     NA     0.0303         7.36  3.24  1.96 1.01   13.7
#>  8 s(x0)  x0    NA     NA     0.0353         7.34  3.20  1.96 1.07   13.6
#>  9 s(x0)  x0    NA     NA     0.0403         7.32  3.15  1.96 1.14   13.5
#> 10 s(x0)  x0    NA     NA     0.0453         7.29  3.10  1.96 1.21   13.4
#> # ... with 790 more rows

## derivatives for a selected smooth
derivatives(mod, type = "central", term = "s(x1)")
#> # A tibble: 200 x 10
#>    smooth var   by_var fs_var     data derivative    se  crit lower upper
#>    <chr>  <chr> <chr>  <chr>     <dbl>      <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 s(x1)  x1    NA     NA     0.000405     -0.907  3.12  1.96 -7.02  5.20
#>  2 s(x1)  x1    NA     NA     0.00541      -0.907  3.12  1.96 -7.02  5.20
#>  3 s(x1)  x1    NA     NA     0.0104       -0.906  3.11  1.96 -7.01  5.20
#>  4 s(x1)  x1    NA     NA     0.0154       -0.903  3.11  1.96 -6.99  5.19
#>  5 s(x1)  x1    NA     NA     0.0204       -0.898  3.10  1.96 -6.97  5.17
#>  6 s(x1)  x1    NA     NA     0.0254       -0.891  3.08  1.96 -6.93  5.15
#>  7 s(x1)  x1    NA     NA     0.0304       -0.880  3.06  1.96 -6.88  5.12
#>  8 s(x1)  x1    NA     NA     0.0354       -0.867  3.04  1.96 -6.82  5.08
#>  9 s(x1)  x1    NA     NA     0.0404       -0.850  3.00  1.96 -6.74  5.04
#> 10 s(x1)  x1    NA     NA     0.0454       -0.829  2.96  1.96 -6.64  4.98
#> # ... with 190 more rows
## or via a partial match
derivatives(mod, type = "central", term = "x1", partial_match = TRUE)
#> # A tibble: 200 x 10
#>    smooth var   by_var fs_var     data derivative    se  crit lower upper
#>    <chr>  <chr> <chr>  <chr>     <dbl>      <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 s(x1)  x1    NA     NA     0.000405     -0.907  3.12  1.96 -7.02  5.20
#>  2 s(x1)  x1    NA     NA     0.00541      -0.907  3.12  1.96 -7.02  5.20
#>  3 s(x1)  x1    NA     NA     0.0104       -0.906  3.11  1.96 -7.01  5.20
#>  4 s(x1)  x1    NA     NA     0.0154       -0.903  3.11  1.96 -6.99  5.19
#>  5 s(x1)  x1    NA     NA     0.0204       -0.898  3.10  1.96 -6.97  5.17
#>  6 s(x1)  x1    NA     NA     0.0254       -0.891  3.08  1.96 -6.93  5.15
#>  7 s(x1)  x1    NA     NA     0.0304       -0.880  3.06  1.96 -6.88  5.12
#>  8 s(x1)  x1    NA     NA     0.0354       -0.867  3.04  1.96 -6.82  5.08
#>  9 s(x1)  x1    NA     NA     0.0404       -0.850  3.00  1.96 -6.74  5.04
#> 10 s(x1)  x1    NA     NA     0.0454       -0.829  2.96  1.96 -6.64  4.98
#> # ... with 190 more rows
# \dontshow{
options(op)
# }