Differences of factor smooth interactions

Differences of factor smooth interactions

difference_smooths(model, ...)

# S3 method for gam
difference_smooths(
  model,
  smooth,
  n = 100,
  ci_level = 0.95,
  newdata = NULL,
  unconditional = FALSE,
  frequentist = FALSE,
  ...
)

Arguments

model
A fitted model.

...
arguments passed to other methods.

smooth
character; which smooth to compute differences for.

n
numeric; the number of points at which to evaluate the difference between pairs of smooths.

ci_level
numeric between 0 and 1; the coverage of credible interval.

newdata
data frame of locations at which to evaluate the difference between smooths.

unconditional
logical; account for smoothness selection in the model?

frequentist
logical; use the frequentist covariance matrix?

Examples

load_mgcv()
set.seed(42)
df <- data_sim("eg4")
m <- gam(y ~ fac + s(x2, by = fac) + s(x0), data = df)

difference_smooths(m, smooth = "s(x2)")

#> # A tibble: 300 x 9
#>    smooth by    level_1 level_2  diff    se     lower upper      x2
#>    <chr>  <chr> <chr>   <chr>   <dbl> <dbl>     <dbl> <dbl>   <dbl>
#>  1 s(x2)  fac   1       2       0.578 0.579 -0.557     1.71 0.00359
#>  2 s(x2)  fac   1       2       0.654 0.544 -0.411     1.72 0.0136 
#>  3 s(x2)  fac   1       2       0.731 0.510 -0.269     1.73 0.0237 
#>  4 s(x2)  fac   1       2       0.807 0.479 -0.132     1.75 0.0338 
#>  5 s(x2)  fac   1       2       0.884 0.451 -0.000608  1.77 0.0438 
#>  6 s(x2)  fac   1       2       0.960 0.426  0.125     1.79 0.0539 
#>  7 s(x2)  fac   1       2       1.04  0.404  0.243     1.83 0.0639 
#>  8 s(x2)  fac   1       2       1.11  0.385  0.354     1.87 0.0740 
#>  9 s(x2)  fac   1       2       1.18  0.370  0.459     1.91 0.0841 
#> 10 s(x2)  fac   1       2       1.26  0.357  0.557     1.96 0.0941 
#> # … with 290 more rows