Evaluate a smooth

Evaluate a smooth at a grid of evenly spaced value over the range of the covariate associated with the smooth. Alternatively, a set of points at which the smooth should be evaluated can be supplied.

evaluate_smooth(object, ...)

# S3 method for gam
evaluate_smooth(object, smooth, n = 100, newdata = NULL,
  unconditional = FALSE, inc_mean = FALSE, dist = 0.1, ...)

# S3 method for gamm
evaluate_smooth(object, ...)

evaluate_parametric_term(object, ...)

# S3 method for gam
evaluate_parametric_term(object, term,
  unconditional = FALSE, ...)

Arguments

object
an object of class "gam" or "gamm".

...
arguments passed to other methods.

smooth
character; a single smooth to evaluate.

n
numeric; the number of points over the range of the covariate at which to evaluate the smooth.

newdata
a vector or data frame of points at which to evaluate the smooth.

unconditional
logical; should confidence intervals include the uncertainty due to smoothness selection? If TRUE, the corrected Bayesian covariance matrix will be used.

inc_mean
logical; should the uncertainty in the model constant term be included in the standard error of the evaluate values of the smooth? Currently not implemented.

dist
numeric; if greater than 0, this is used to determine when a location is too far from data to be plotted when plotting 2-D smooths. The data are scaled into the unit square before deciding what to exclude, and dist is a distance within the unit square. See mgcv::exclude.too.far() for further details.

term
character; which parametric term whose effects are evaulated

Value

A data frame, which is of class "evaluated_1d_smooth" or evaluated_2d_smooth, which inherit from classes "evaluated_smooth" and "data.frame".

Examples

library("mgcv")
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

head(evaluate_smooth(m1, "s(x1)"))

#>   smooth           x1       est        se
#> 1  s(x1) 0.0006632213 -1.944733 0.3134611
#> 2  s(x1) 0.0107501579 -1.930377 0.2957587
#> 3  s(x1) 0.0208370945 -1.916014 0.2788680
#> 4  s(x1) 0.0309240311 -1.901632 0.2629466
#> 5  s(x1) 0.0410109677 -1.887206 0.2481575
#> 6  s(x1) 0.0510979043 -1.872698 0.2346391

## 2d example
set.seed(2)
dat <- gamSim(2, n = 4000, dist = "normal", scale = 1)

#> Bivariate smoothing example

m2 <- gam(y ~ s(x, z, k = 30), data = dat$data, method = "REML")

head(evaluate_smooth(m2, "s(x,z)", n = 100))

#>   smooth            x            z       est        se
#> 1 s(x,z) 2.298248e-05 4.321919e-05 0.1226246 0.1149179
#> 2 s(x,z) 1.012240e-02 4.321919e-05 0.1218180 0.1124405
#> 3 s(x,z) 2.022182e-02 4.321919e-05 0.1208805 0.1099417
#> 4 s(x,z) 3.032124e-02 4.321919e-05 0.1198083 0.1074268
#> 5 s(x,z) 4.042066e-02 4.321919e-05 0.1185977 0.1049027
#> 6 s(x,z) 5.052008e-02 4.321919e-05 0.1172456 0.1023753