Evaluate a smooth

Evaluate a smooth at a grid of evenly spaced value over the range of the covariate associated with the smooth. Alternatively, a set of points at which the smooth should be evaluated can be supplied.

Usage

evaluate_smooth(object, ...)

# S3 method for gam
evaluate_smooth(
  object,
  smooth,
  n = 100,
  newdata = NULL,
  unconditional = FALSE,
  overall_uncertainty = TRUE,
  dist = 0.1,
  ...
)

# S3 method for gamm
evaluate_smooth(object, ...)

# S3 method for list
evaluate_smooth(object, ...)

Arguments

object

an object of class "gam" or "gamm".

...

arguments passed to other methods.

smooth

character; a single smooth to evaluate.

n

numeric; the number of points over the range of the covariate at which to evaluate the smooth.

newdata

a vector or data frame of points at which to evaluate the smooth.

unconditional

logical; should confidence intervals include the uncertainty due to smoothness selection? If TRUE, the corrected Bayesian covariance matrix will be used.

overall_uncertainty

logical; should the uncertainty in the model constant term be included in the standard error of the evaluate values of the smooth?

dist

numeric; if greater than 0, this is used to determine when a location is too far from data to be plotted when plotting 2-D smooths. The data are scaled into the unit square before deciding what to exclude, and dist is a distance within the unit square. See mgcv::exclude.too.far() for further details.

Value

A data frame, which is of class "evaluated_1d_smooth" or evaluated_2d_smooth, which inherit from classes "evaluated_smooth"

and "data.frame".

Details

evaluate_smooth() is deprecated in favour of smooth_estimates(), which provides a cleaner way to evaluate a smooth over a range of covariate values. smooth_estimates() can handle a much wider range of models than evaluate_smooth() is capable of and smooth_estimates() is much easier to extend to handle new smooth types.

Most code that uses evaluate_smooth() should work simply by changing the function call to smooth_estimates(). However, there are some differences:

the newdata argument becomes data

Consider evaluate_smooth() to be soft-deprecated; its use is discouraged and it may be removed at a later date if it becomes difficult to maintain the current functionality, but there are no intentions of removing it from gratia unless that situation arises.

Examples

load_mgcv()
# \dontshow{
op <- options(cli.unicode = FALSE, pillar.sigfig = 2)
# }
dat <- data_sim("eg1", n = 500, dist = "normal", scale = 1, seed = 2)
m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

evaluate_smooth(m1, "s(x1)")
#> # A tibble: 100 x 5
#>    smooth by_variable     x1   est    se
#>    <chr>  <fct>        <dbl> <dbl> <dbl>
#>  1 s(x1)  NA          0.0025  -2.1 0.17 
#>  2 s(x1)  NA          0.013   -2.1 0.15 
#>  3 s(x1)  NA          0.023   -2.1 0.14 
#>  4 s(x1)  NA          0.033   -2.0 0.13 
#>  5 s(x1)  NA          0.043   -2.0 0.12 
#>  6 s(x1)  NA          0.053   -2.0 0.11 
#>  7 s(x1)  NA          0.063   -2.0 0.11 
#>  8 s(x1)  NA          0.073   -2.0 0.10 
#>  9 s(x1)  NA          0.083   -2.0 0.10 
#> 10 s(x1)  NA          0.093   -2.0 0.097
#> # ... with 90 more rows

## 2d example
dat <- data_sim("eg2", n = 1000, dist = "normal", scale = 1, seed = 2)
m2 <- gam(y ~ s(x, z, k = 30), data = dat, method = "REML")

evaluate_smooth(m2, "s(x,z)", n = 50)
#> # A tibble: 2,500 x 6
#>    smooth by_variable       x       z   est    se
#>    <chr>  <fct>         <dbl>   <dbl> <dbl> <dbl>
#>  1 s(x,z) NA          0.00066 0.00076 0.088  0.19
#>  2 s(x,z) NA          0.021   0.00076 0.097  0.18
#>  3 s(x,z) NA          0.041   0.00076 0.11   0.18
#>  4 s(x,z) NA          0.062   0.00076 0.12   0.17
#>  5 s(x,z) NA          0.082   0.00076 0.13   0.16
#>  6 s(x,z) NA          0.10    0.00076 0.14   0.16
#>  7 s(x,z) NA          0.12    0.00076 0.14   0.15
#>  8 s(x,z) NA          0.14    0.00076 0.15   0.15
#>  9 s(x,z) NA          0.16    0.00076 0.16   0.14
#> 10 s(x,z) NA          0.18    0.00076 0.17   0.14
#> # ... with 2,490 more rows
options(op)