Draw fitted values from the posterior distribution

Expectations (fitted values) of the response drawn from the posterior distribution of fitted model using a Gaussian approximation to the posterior.

fitted_samples(model, ...)

# S3 method for gam
fitted_samples(
  model,
  n = 1,
  newdata,
  seed,
  scale = c("response", "linear_predictor"),
  method = c("gaussian", "mh", "inla"),
  freq = FALSE,
  unconditional = FALSE,
  ncores = 1L,
  ...
)

Arguments

model
a fitted model of the supported types

...
arguments passed to other methods. For fitted_samples(), these are passed on to predict.gam().

n
numeric; the number of posterior samples to return.

newdata
data frame; new observations at which the posterior draws from the model should be evaluated. If not supplied, the data used to fit the model will be used for newdata, if available in model.

seed
numeric; a random seed for the simulations.

scale
character;

method
character; the method used to generate samples from the posterior distribution of the model. "gaussian", the default, uses a Gaussian approximation to the posterior. "mh" uses a simple Metropolis Hastings sampler, while "inla" uses a variant of Integrated Nested Laplace Approximation due to Wood (2019). Currently, the only available option is "gaussian".

freq
logical; TRUE to use the frequentist covariance matrix of the parameter estimators, FALSE to use the Bayesian posterior covariance matrix of the parameters.

unconditional
logical; if TRUE (and freq == FALSE) then the Bayesian smoothing parameter uncertainty corrected covariance matrix is used, if available.

ncores
number of cores for generating random variables from a multivariate normal distribution. Passed to mvnfast::rmvn(). Parallelization will take place only if OpenMP is supported (but appears to work on Windows with current R).

Value

A tibble (data frame) with 3 columns containing the posterior predicted values in long format. The columns are

row (integer) the row of newdata that each posterior draw relates to,

draw (integer) an index, in range 1:n, indicating which draw each row relates to,

response (numeric) the predicted response for the indicated row of newdata.

References

Wood, S.N., (2020). Simplified integrated nested Laplace approximation. Biometrika 107, 223--230. doi: 10.1093/biomet/asz044

Author

Gavin L. Simpson

Examples

load_mgcv()
# \dontshow{
set.seed(2)
op <- options(cli.unicode = FALSE)
# }
dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
#> Gu & Wahba 4 term additive model
m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

fitted_samples(m1, n = 5, seed = 42)
#> # A tibble: 5,000 x 3
#>      row  draw fitted
#>    <int> <int>  <dbl>
#>  1     1     1   6.24
#>  2     2     1   5.12
#>  3     3     1   7.06
#>  4     4     1   7.56
#>  5     5     1   9.20
#>  6     6     1   7.94
#>  7     7     1   6.13
#>  8     8     1   7.59
#>  9     9     1  13.9 
#> 10    10     1  12.7 
#> # ... with 4,990 more rows
options(op)