Generate fitted values from a estimated GAM

Generate fitted values from a estimated GAM

fitted_values(object, ...)

# S3 method for gam
fitted_values(
  object,
  data = NULL,
  scale = c("response", "link", "linear predictor"),
  ci_level = 0.95,
  ...
)

Arguments

object

a fitted model. Currently only models fitted by mgcv::gam() and mgcv::bam() are supported.

...

arguments passed to mgcv::predict.gam(). Note that type, newdata, and se.fit are already used and passed on to mgcv::predict.gam().

data

optional data frame of covariate values for which fitted values are to be returned.

scale

character; what scale should the fitted values be returned on? "linear predictor" is a synonym for "link" if you prefer that terminology.

ci_level

numeric; a value between 0 and 1 indicating the coverage of the credible interval.

Value

A tibble (data frame) whose first m columns contain either the data used to fit the model (if data was NULL), or the variables supplied to data. Four further columns are added:

fitted: the fitted values on the specified scale,

se: the standard error of the fitted values (always on the link scale),

lower, upper: the limits of the credible interval on the fitted values, on the specified scale.

Note

Regardless of the scale on which the fitted values are returned, the se component of the returned object is on the link (linear predictor) scale, not the response scale.

Examples

load_mgcv()
# \dontshow{
op <- options(cli.unicode = FALSE, digits = 6)
# }
sim_df <- data_sim("eg1", n = 400, dist = "normal", scale = 2, seed = 2)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = sim_df, method = "REML")
fv <- fitted_values(m)
fv
#> # A tibble: 400 x 8
#>       x0     x1      x2    x3 fitted    se lower upper
#>    <dbl>  <dbl>   <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 0.185 0.617  0.415   0.132   8.74 0.355  8.04  9.43
#>  2 0.702 0.569  0.531   0.365   7.63 0.338  6.96  8.29
#>  3 0.573 0.154  0.00325 0.455   3.12 0.592  1.96  4.28
#>  4 0.168 0.0348 0.252   0.537  11.1  0.402 10.3  11.9 
#>  5 0.944 0.998  0.155   0.185  14.1  0.453 13.2  14.9 
#>  6 0.943 0.836  0.879   0.449   6.13 0.365  5.42  6.85
#>  7 0.129 0.587  0.204   0.257  12.5  0.356 11.8  13.2 
#>  8 0.833 0.339  0.584   0.618   6.25 0.345  5.58  6.93
#>  9 0.468 0.167  0.804   0.881   4.21 0.372  3.49  4.94
#> 10 0.550 0.807  0.265   0.318  15.5  0.370 14.8  16.3 
#> # ... with 390 more rows
options(op)