Return the linear prediction matrix of a fitted GAM

lp_matrix() is a wrapper to predict(..., type = "lpmatrix") for returning the linear predictor matrix for the model training data (when data = NULL), or user-specified data values supplied via data.

Usage

lp_matrix(model, ...)

# S3 method for gam
lp_matrix(model, data = NULL, ...)

Arguments

model

a fitted model

...

arguments passed to other methods and predict methods including mgcv::predict.gam() and mgcv::predict.bam()

data

a data frame of values at which to return the linear prediction matrix.

Value

The linear prediction matrix is returned as a matrix. The object returned is of class "lp_matrix", which inherits from classes "matrix"

and "array". The special class allows the printing of the matrix to be controlled, which we do by printing the matrix as a tibble.

Details

The linear prediction matrix \(\mathbf{X}_p\) is a matrix that maps values of parameters \(\hat{\mathbf{\beta}}_p\) to values on the linear predictor of the model \(\hat{\eta}_p = \mathbf{X}_p \hat{\mathbf{\beta}}_p\). \(\mathbf{X}_p\) is the model matrix where spline covariates have been replaced by the values of the basis functions evaluated at the respective covariates. Parametric covariates are also included.

Examples

load_mgcv()
op <- options(digits = 3, cli.unicode = FALSE)
df <- data_sim("eg1", seed = 1)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df)

# linear prediction matrix for observed data
xp <- lp_matrix(m)
xp
#> Linear prediction matrix (400 x 37)
#>   `(Intercept)` s(x0).~1 s(x0)~2 s(x0)~3 s(x0)~4 s(x0)~5 s(x0)~6 s(x0)~7 s(x0)~8
#>   <matrix>      <matrix> <matri> <matri> <matri> <matri> <matri> <matri> <matri>
#> 1 1              0.961    0.227   0.7058 -0.135   0.457  -0.146~  0.6209  0.2016
#> 2 1              0.651   -0.241   0.0684 -0.308   0.394  -0.009~  0.3551  0.4182
#> 3 1             -0.385   -0.549   0.0660 -0.204  -0.416  -0.247~ -0.0349  0.4622
#> 4 1             -1.273    0.156  -1.5344  0.222  -1.598   0.197~ -1.5836 -0.7479
#> 5 1              1.051    0.420   1.1138  0.214   0.893   0.035~  0.7760 -0.0289
#> # ... with 395 more rows, 28 more variables: `s(x0).9` <matrix>, ...

# the object `xp` *is* a matrix
class(xp)
#> [1] "lp_matrix" "matrix"    "array"    
# but we print like a tibble to avoid spamming the R console

# linear predictor matrix for new data set
ds <- data_slice(m, x2 = evenly(x2))
xp <- lp_matrix(m, data = ds)
xp
#> Linear prediction matrix (100 x 37)
#>   `(Intercept)` s(x0).~1 s(x0)~2 s(x0)~3 s(x0)~4 s(x0)~5 s(x0)~6 s(x0)~7 s(x0)~8
#>   <matrix>      <matrix> <matri> <matri> <matri> <matri> <matri> <matri> <matri>
#> 1 1             0.17     -0.542  -0.0371 -0.0534 0.144   -0.353  -0.0572 0.518  
#> 2 1             0.17     -0.542  -0.0371 -0.0534 0.144   -0.353  -0.0572 0.518  
#> 3 1             0.17     -0.542  -0.0371 -0.0534 0.144   -0.353  -0.0572 0.518  
#> 4 1             0.17     -0.542  -0.0371 -0.0534 0.144   -0.353  -0.0572 0.518  
#> 5 1             0.17     -0.542  -0.0371 -0.0534 0.144   -0.353  -0.0572 0.518  
#> # ... with 95 more rows, 28 more variables: `s(x0).9` <matrix>, ...
options(op)