Estimated values for parametric model terms

Estimated values for parametric model terms

parametric_effects(object, ...)

# S3 method for gam
parametric_effects(
  object,
  term = NULL,
  data = NULL,
  unconditional = FALSE,
  unnest = TRUE,
  ci_level = 0.95,
  envir = environment(formula(object)),
  ...
)

Arguments

object

a fitted model object.

...

arguments passed to other methods.

term

character; which parametric term whose effects are evaluated.

data

a data frame of covariate values used to fit the model. Can be missing, in which case the data will be recovered from the model.

unconditional

logical; should confidence intervals include the uncertainty due to smoothness selection? If TRUE, the corrected Bayesian covariance matrix will be used.

unnest

logical; unnest the smooth objects?

ci_level

numeric; the coverage required for the confidence interval. Currently ignored.

envir

an environment to look up the data within.