Partial residuals

Partial residuals

Usage

partial_residuals(object, ...)

# S3 method for gam
partial_residuals(object, select = NULL, partial_match = FALSE, ...)

Arguments

object

an R object, typically a model. Currently only objects of class "gam" (or that inherit from that class) are supported.

...

arguments passed to other methods.

select

character, logical, or numeric; which smooths to plot. If NULL, the default, then all model smooths are drawn. Numeric select indexes the smooths in the order they are specified in the formula and stored in object. Character select matches the labels for smooths as shown for example in the output from summary(object). Logical select operates as per numeric select in the order that smooths are stored.

partial_match

logical; should smooths be selected by partial matches with select? If TRUE, select can only be a single string to match against.

Examples

# \dontshow{
set.seed(1)
op <- options(digits = 4, cli.unicode = FALSE)
# }
## load mgcv
load_mgcv()

## example data - Gu & Wabha four term model
df <- data_sim("eg1", n = 400, seed = 42)
## fit the model
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = 'REML')

## extract partial residuals
partial_residuals(m)
#> # A tibble: 400 x 4
#>    `s(x0)` `s(x1)` `s(x2)` `s(x3)`
#>      <dbl>   <dbl>   <dbl>   <dbl>
#>  1  -0.353 -1.32    -2.18   0.608 
#>  2  -0.123  0.501   -1.77   0.961 
#>  3   1.43   1.52     5.61   0.991 
#>  4  -1.11  -1.70    -0.888 -0.659 
#>  5  -2.12  -0.0138  -2.73  -3.01  
#>  6   1.25  -1.22     3.91   0.0728
#>  7  -0.522  3.02    -0.820 -1.02  
#>  8   1.40   0.218    7.06   1.90  
#>  9   2.80   0.497    7.33   2.50  
#> 10   1.15  -0.227    0.720  0.744 
#> # ... with 390 more rows

## and for a select term
partial_residuals(m, select = "s(x2)")
#> # A tibble: 400 x 1
#>    `s(x2)`
#>      <dbl>
#>  1  -2.18 
#>  2  -1.77 
#>  3   5.61 
#>  4  -0.888
#>  5  -2.73 
#>  6   3.91 
#>  7  -0.820
#>  8   7.06 
#>  9   7.33 
#> 10   0.720
#> # ... with 390 more rows

## or with partial matching
partial_residuals(m, select = "x", partial_match = TRUE) # returns all
#> # A tibble: 400 x 4
#>    `s(x0)` `s(x1)` `s(x2)` `s(x3)`
#>      <dbl>   <dbl>   <dbl>   <dbl>
#>  1  -0.353 -1.32    -2.18   0.608 
#>  2  -0.123  0.501   -1.77   0.961 
#>  3   1.43   1.52     5.61   0.991 
#>  4  -1.11  -1.70    -0.888 -0.659 
#>  5  -2.12  -0.0138  -2.73  -3.01  
#>  6   1.25  -1.22     3.91   0.0728
#>  7  -0.522  3.02    -0.820 -1.02  
#>  8   1.40   0.218    7.06   1.90  
#>  9   2.80   0.497    7.33   2.50  
#> 10   1.15  -0.227    0.720  0.744 
#> # ... with 390 more rows
# \dontshow{
options(op)
# }