Extract and tidy penalty matrices

Extract and tidy penalty matrices

penalty(object, ...)

# S3 method for gam
penalty(object, smooth = NULL, rescale = FALSE, ...)

# S3 method for mgcv.smooth
penalty(object, rescale = FALSE, ...)

# S3 method for tensor.smooth
penalty(object, margins = FALSE, ...)

# S3 method for t2.smooth
penalty(object, margins = FALSE, ...)

Arguments

object
a fitted GAM or a smooth.

...
additional arguments passed to methods.

smooth
character; vector of smooths to extract penalty matrices for. If NULL, penalty matrices for all smooths in object are extracted.

rescale
logical; by default, mgcv will scale the penalty matrix for better performance in mgcv::gamm(). If rescale is TRUE, this scaling will be undone to put the penalty matrix back on the original scale.

margins
logical; extract the penalty matrices for the tensor product or the margainl smooths of the tensor product?

Author

Gavin L. Simpson

Examples

load_mgcv()
dat <- data_sim("eg4", n = 400, seed = 42)
m <- gam(y ~ s(x0) + s(x1) + s(x2, by = fac),
         data = dat, method = "REML")
penalty(m)

#> # A tibble: 405 x 6
#>    smooth type  penalty row   col       value
#>    <chr>  <chr> <chr>   <chr> <chr>     <dbl>
#>  1 s(x0)  TPRS  s(x0)   f1    f1     9.81e+ 0
#>  2 s(x0)  TPRS  s(x0)   f1    f2    -1.45e+ 0
#>  3 s(x0)  TPRS  s(x0)   f1    f3    -5.00e+ 0
#>  4 s(x0)  TPRS  s(x0)   f1    f4    -1.34e+ 0
#>  5 s(x0)  TPRS  s(x0)   f1    f5    -6.24e+ 0
#>  6 s(x0)  TPRS  s(x0)   f1    f6     3.90e+ 0
#>  7 s(x0)  TPRS  s(x0)   f1    f7    -7.74e+ 0
#>  8 s(x0)  TPRS  s(x0)   f1    f8    -1.79e+ 0
#>  9 s(x0)  TPRS  s(x0)   f1    f9     1.17e-16
#> 10 s(x0)  TPRS  s(x0)   f2    f1    -1.45e+ 0
#> # … with 395 more rows

# for a specific smooth
penalty(m, smooth = "s(x2):fac1")

#> # A tibble: 81 x 6
#>    smooth     type  penalty    row   col       value
#>    <chr>      <chr> <chr>      <chr> <chr>     <dbl>
#>  1 s(x2):fac1 TPRS  s(x2):fac1 f1    f1     9.94e+ 0
#>  2 s(x2):fac1 TPRS  s(x2):fac1 f1    f2     1.40e+ 0
#>  3 s(x2):fac1 TPRS  s(x2):fac1 f1    f3     4.79e+ 0
#>  4 s(x2):fac1 TPRS  s(x2):fac1 f1    f4     1.13e+ 0
#>  5 s(x2):fac1 TPRS  s(x2):fac1 f1    f5     5.84e+ 0
#>  6 s(x2):fac1 TPRS  s(x2):fac1 f1    f6    -4.06e+ 0
#>  7 s(x2):fac1 TPRS  s(x2):fac1 f1    f7     8.60e+ 0
#>  8 s(x2):fac1 TPRS  s(x2):fac1 f1    f8     1.44e+ 0
#>  9 s(x2):fac1 TPRS  s(x2):fac1 f1    f9    -8.82e-17
#> 10 s(x2):fac1 TPRS  s(x2):fac1 f2    f1     1.40e+ 0
#> # … with 71 more rows