Draw predicted values from the posterior distribution

Predicted values of the response drawn from the posterior distribution of fitted model, created via simulate() (e.g. simulate.gam()) and returned in a tidy, long, format.

# S3 method for gam
fitted_samples(model, n, newdata, seed,
  scale = c("response", "linear_predictor"), freq = FALSE,
  unconditional = FALSE, ncores = 1L, ...)

predicted_samples(model, ...)

# S3 method for gam
predicted_samples(model, n = 1, newdata = NULL,
  seed = NULL, freq = FALSE, unconditional = FALSE, weights = NULL,
  ...)

Arguments

model
a fitted model of the supported types

n
numeric; the number of posterior samples to return.

newdata
data frame; new observations at which the posterior draws from the model should be evaluated. If not supplied, the data used to fit the model will be used for newdata, if available in model.

seed
numeric; a random seed for the simulations.

scale
character;

freq
logical; TRUE to return the frequentist covariance matrix of the parameter estimators, FALSE to return the Bayesian posterior covariance matrix of the parameters.

unconditional
logical; if TRUE (and freq == FALSE) then the Bayesian smoothing parameter uncertainty corrected covariance matrix is returned, if available.

ncores
number of cores for generating random variables from a multivariate normal distribution. Passed to mvnfast::rmvn(). Parallelization will take place only if OpenMP is supported (but appears to work on Windows with current R).

...
arguments passed to other methods

weights
numeric; a vector of prior weights. If newdata is null then defaults to object[["prior.weights"]], otherwise a vector of ones.

Value

A tibble (data frame) with 3 columns containing the posterior predicted values in long format. The columns are

row (integet) the row of newdata that each posterior draw relates to,

draw (integer) an index, in range 1:n, indicating which draw each row relates to,

response (numeric) the predicted response for the indicated row of newdata.

Examples

suppressPackageStartupMessages(library("mgcv"))
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

fitted_samples(m1, n = 5, seed = 42)

#> # A tibble: 2,000 x 3
#>      row  draw fitted
#>    <int> <int>  <dbl>
#>  1     1     1   8.72
#>  2     2     1   7.79
#>  3     3     1   2.94
#>  4     4     1  10.5 
#>  5     5     1  13.9 
#>  6     6     1   5.89
#>  7     7     1  12.0 
#>  8     8     1   6.14
#>  9     9     1   4.44
#> 10    10     1  15.2 
#> # … with 1,990 more rows

suppressPackageStartupMessages(library("mgcv"))
set.seed(2)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

predicted_samples(m1, n = 5, seed = 42)

#> # A tibble: 2,000 x 3
#>      row  draw response
#>    <int> <int>    <dbl>
#>  1     1     1    11.4 
#>  2     2     1     6.51
#>  3     3     1     3.84
#>  4     4     1    12.4 
#>  5     5     1    14.9 
#>  6     6     1     5.92
#>  7     7     1    15.5 
#>  8     8     1     6.07
#>  9     9     1     8.20
#> 10    10     1    15.4 
#> # … with 1,990 more rows