Draw new response values from the conditional distribution of the response

Predicted values of the response (new response data) are drawn from the fitted model, created via simulate() (e.g. simulate.gam()) and returned in a tidy, long, format. These predicted values do not include the uncertainty in the estimated model; they are simply draws from the conditional distribution of the response.

predicted_samples(model, ...)

# S3 method for gam
predicted_samples(
  model,
  n = 1,
  newdata = NULL,
  seed = NULL,
  weights = NULL,
  ...
)

Arguments

model
a fitted model of the supported types

...
arguments passed to other methods. For fitted_samples(), these are passed on to predict.gam().

n
numeric; the number of posterior samples to return.

newdata
data frame; new observations at which the posterior draws from the model should be evaluated. If not supplied, the data used to fit the model will be used for newdata, if available in model.

seed
numeric; a random seed for the simulations.

weights
numeric; a vector of prior weights. If newdata is null then defaults to object[["prior.weights"]], otherwise a vector of ones.

Value

A tibble (data frame) with 3 columns containing the posterior predicted values in long format. The columns are

row (integer) the row of newdata that each posterior draw relates to,

draw (integer) an index, in range 1:n, indicating which draw each row relates to,

response (numeric) the predicted response for the indicated row of newdata.

Author

Gavin L. Simpson

Examples

load_mgcv()
# \dontshow{
set.seed(2)
op <- options(cli.unicode = FALSE)
# }
dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
#> Gu & Wahba 4 term additive model
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

predicted_samples(m, n = 5, seed = 42)
#> # A tibble: 5,000 x 3
#>      row  draw response
#>    <int> <int>    <dbl>
#>  1     1     1     8.93
#>  2     2     1     4.23
#>  3     3     1     7.71
#>  4     4     1     8.51
#>  5     5     1    10.1 
#>  6     6     1     8.20
#>  7     7     1     8.95
#>  8     8     1     7.20
#>  9     9     1    18.1 
#> 10    10     1    12.7 
#> # ... with 4,990 more rows

## Can pass arguments to predict.gam()
newd <- data.frame(x0 = runif(10), x1 = runif(10), x2 = runif(10),
                   x3 = runif(10))

## Exclude s(x2)
predicted_samples(m, n = 5, newd, exclude = "s(x2)", seed = 25)
#> # A tibble: 50 x 3
#>      row  draw response
#>    <int> <int>    <dbl>
#>  1     1     1     4.16
#>  2     2     1     7.32
#>  3     3     1     8.69
#>  4     4     1     6.36
#>  5     5     1     4.14
#>  6     6     1     8.28
#>  7     7     1    13.2 
#>  8     8     1    13.1 
#>  9     9     1     9.10
#> 10    10     1     6.27
#> # ... with 40 more rows

## Exclude s(x1)
predicted_samples(m, n = 5, newd, exclude = "s(x1)", seed = 25)
#> # A tibble: 50 x 3
#>      row  draw response
#>    <int> <int>    <dbl>
#>  1     1     1     6.00
#>  2     2     1    11.7 
#>  3     3     1     2.05
#>  4     4     1     9.03
#>  5     5     1     3.53
#>  6     6     1     8.29
#>  7     7     1    16.8 
#>  8     8     1    12.4 
#>  9     9     1    10.1 
#> 10    10     1     6.68
#> # ... with 40 more rows

## Select which terms --- result same as previous
predicted_samples(m, n = 5, newd, seed = 25,
                  terms = c("s(x0)", "s(x2)", "s(x3)"))
#> # A tibble: 50 x 3
#>      row  draw response
#>    <int> <int>    <dbl>
#>  1     1     1     6.00
#>  2     2     1    11.7 
#>  3     3     1     2.05
#>  4     4     1     9.03
#>  5     5     1     3.53
#>  6     6     1     8.29
#>  7     7     1    16.8 
#>  8     8     1    12.4 
#>  9     9     1    10.1 
#> 10    10     1     6.68
#> # ... with 40 more rows
options(op)