New evaluate_smooth() alike

New evaluate_smooth() alike

smooth_estimates(object, ...)

# S3 method for gam
smooth_estimates(
  object,
  smooth = NULL,
  n = 100,
  newdata = NULL,
  unconditional = FALSE,
  overall_uncertainty = TRUE,
  dist = 0.1,
  ...
)

Arguments

object
an object of class "gam" or "gamm".

...
arguments passed to other methods.

smooth
character; a single smooth to evaluate.

n
numeric; the number of points over the range of the covariate at which to evaluate the smooth.

newdata
a vector or data frame of points at which to evaluate the smooth.

unconditional
logical; should confidence intervals include the uncertainty due to smoothness selection? If TRUE, the corrected Bayesian covariance matrix will be used.

overall_uncertainty
logical; should the uncertainty in the model constant term be included in the standard error of the evaluate values of the smooth?

dist
numeric; if greater than 0, this is used to determine when a location is too far from data to be plotted when plotting 2-D smooths. The data are scaled into the unit square before deciding what to exclude, and dist is a distance within the unit square. See mgcv::exclude.too.far() for further details.

Value

A data frame, which is of class "evaluated_1d_smooth" or evaluated_2d_smooth, which inherit from classes "evaluated_smooth" and "data.frame".

Examples

load_mgcv()
# \dontshow{
set.seed(2)
op <- options(cli.unicode = FALSE, digits = 6)
# }
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")

## evaluate all smooths
smooth_estimates(m1)

#> # A tibble: 400 x 9
#>    smooth type  by       est    se      x0    x1    x2    x3
#>    <chr>  <chr> <chr>  <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 s(x0)  TPRS  NA    -0.967 0.316 0.00711    NA    NA    NA
#>  2 s(x0)  TPRS  NA    -0.925 0.297 0.0171     NA    NA    NA
#>  3 s(x0)  TPRS  NA    -0.884 0.279 0.0271     NA    NA    NA
#>  4 s(x0)  TPRS  NA    -0.843 0.263 0.0371     NA    NA    NA
#>  5 s(x0)  TPRS  NA    -0.802 0.247 0.0471     NA    NA    NA
#>  6 s(x0)  TPRS  NA    -0.761 0.234 0.0571     NA    NA    NA
#>  7 s(x0)  TPRS  NA    -0.719 0.222 0.0671     NA    NA    NA
#>  8 s(x0)  TPRS  NA    -0.678 0.211 0.0772     NA    NA    NA
#>  9 s(x0)  TPRS  NA    -0.636 0.202 0.0872     NA    NA    NA
#> 10 s(x0)  TPRS  NA    -0.595 0.195 0.0972     NA    NA    NA
#> # ... with 390 more rows

## or selected smooths
smooth_estimates(m1, smooth = c("s(x0)", "s(x1)"))

#> # A tibble: 200 x 7
#>    smooth type  by       est    se      x0    x1
#>    <chr>  <chr> <chr>  <dbl> <dbl>   <dbl> <dbl>
#>  1 s(x0)  TPRS  NA    -0.967 0.316 0.00711    NA
#>  2 s(x0)  TPRS  NA    -0.925 0.297 0.0171     NA
#>  3 s(x0)  TPRS  NA    -0.884 0.279 0.0271     NA
#>  4 s(x0)  TPRS  NA    -0.843 0.263 0.0371     NA
#>  5 s(x0)  TPRS  NA    -0.802 0.247 0.0471     NA
#>  6 s(x0)  TPRS  NA    -0.761 0.234 0.0571     NA
#>  7 s(x0)  TPRS  NA    -0.719 0.222 0.0671     NA
#>  8 s(x0)  TPRS  NA    -0.678 0.211 0.0772     NA
#>  9 s(x0)  TPRS  NA    -0.636 0.202 0.0872     NA
#> 10 s(x0)  TPRS  NA    -0.595 0.195 0.0972     NA
#> # ... with 190 more rows

options(op)